1.0.3

+++ src/main/java/vavi/util/cab/CabFolderInputStream.java
@@ -108,7 +108,7 @@ e.printStackTrace(System.err);
             if ((numread == 0) || (numread == -1)) {
                 iis = readBlock();
             } else {
-                ret += numread;
+                ret += (int) numread;
             }
         }
         return ret;
